extract graphmodel with isomorphism detection into seperate module
