+++ org.eclipse.winery.compliance/src/test/java/org/eclipse/winery/compliance/ToscaGraphIsomorphismTest.java
@@ -38,12 +38,12 @@ import org.eclipse.winery.compliance.checking.ComplianceCheckingException;
 import org.eclipse.winery.compliance.checking.ComplianceRuleChecker;
 import org.eclipse.winery.compliance.checking.ServiceTemplateCheckingResult;
 import org.eclipse.winery.compliance.checking.ServiceTemplateComplianceRuleRuleChecker;
-import org.eclipse.winery.compliance.matching.ToscaComplianceRuleMatcher;
-import org.eclipse.winery.compliance.matching.ToscaIsomorphismMatcher;
-import org.eclipse.winery.compliance.model.ToscaEdge;
-import org.eclipse.winery.compliance.model.ToscaEdgeFactory;
-import org.eclipse.winery.compliance.model.ToscaGraph;
-import org.eclipse.winery.compliance.model.ToscaNode;
+import org.eclipse.winery.compliance.checking.ToscaComplianceRuleMatcher;
+import org.eclipse.winery.topologygraph.matching.ToscaIsomorphismMatcher;
+import org.eclipse.winery.topologygraph.matching.model.ToscaEdge;
+import org.eclipse.winery.topologygraph.matching.model.ToscaEdgeFactory;
+import org.eclipse.winery.topologygraph.matching.model.ToscaGraph;
+import org.eclipse.winery.topologygraph.matching.model.ToscaNode;
 import org.eclipse.winery.model.tosca.TComplianceRule;
 import org.eclipse.winery.model.tosca.TExtensibleElements;
 import org.eclipse.winery.model.tosca.TNodeTemplate;
