+++ org.eclipse.winery.topologymodeler.ui/src/app/navbar/navbar.component.html
@@ -140,6 +140,7 @@
                             placement="left" class="dropdown-item" (click)="exportCsar($event)">Export CSAR
                     </button>
                 </li>
+                <li role="menuitem"><a class="dropdown-item" (click)="checkCompliance()" )>Check Compliance</a></li>
                 <li role="menuitem"><a class="dropdown-item" href="#">About</a></li>
             </ul>
         </div>
