+++ pom.xml
@@ -101,6 +101,7 @@
         <module>org.eclipse.winery.repository.ui</module>
         <module>org.eclipse.winery.topologymodeler.ui</module>
         <module>org.eclipse.winery.topologymodeler</module>
+        <module>org.eclipse.winery.topologygraph</module>
         <module>org.eclipse.winery.workflowmodeler</module>
         <module>org.eclipse.winery.compliance</module>
         <module>org.eclipse.winery.generators.ia</module>
